Lazy load pet action menu content
